clean main code
